+++ src/main/java/com/saiyingge/localVenue/repository/RoleRepository.java
@@ -13,7 +13,7 @@ public interface RoleRepository extends CrudRepository<Role, Long> {
     public Role findRoleByName(String name);
 
     @Query(value = "select * from role where role.id=(select role_id from user_roles where user_id = :id)", nativeQuery = true)
-    public List<Role> findRoleByUserId (int id);
+    public List<Role> findRoleByUserId (long id);
 
 }
